DELIA-57501 - Fix Warnings in RDKServices builds and enable warnings as errors
(cherry picked from commit 6f9731591ea31a6fc6fd48d29b87c5eff53e2a98)

DELIA-57501 - Fix Warnings in RDKServices builds and enable warnings as errors

(cherry picked from commit 35812e50230f2c5e408ed1bb1040c038af4d21eb)

DELIA-57501 - Fix Warnings in RDKServices builds and enable warnings as errors

(cherry picked from commit 53c51351005909c55dbe67af7687e582b28ad769)

DELIA-57501 - Fix Warnings in RDKServices builds and enable warnings as errors

Moving these flags from here to rdkservices_git.bb

(cherry picked from commit 273c8183c8c0fbe0df54b06cbd2954819192f18c)

+++ ActivityMonitor/ActivityMonitor.cpp
@@ -621,7 +621,7 @@ namespace WPEFramework
 
             buf.data()[buf.size() - 1] = 0;
 
-            int pos = 0;
+            unsigned int pos = 0;
             while (pos < buf.size())
             {
                 if (0 == strcmp(buf.data() + pos, CALLSIGN_PARAMETER))
@@ -711,7 +711,7 @@ namespace WPEFramework
                             lastIdx = idx;
                         }
                     }
-                    else if (ppids[idx] == getpid()) // if there is no waylandregistryreceiver.conf, monitoring the children of WPEFramework with "-C <callsign>" parameter
+                    else if (ppids[idx] == (unsigned int)getpid()) // if there is no waylandregistryreceiver.conf, monitoring the children of WPEFramework with "-C <callsign>" parameter
                     {
                         if (pid2callSign.find(pids[idx]) == pid2callSign.end())
                         {
